rename Lib, add Double convert to postgre type
