+++ source/ModelSQLBuilder/ModelBase.cs
@@ -3,9 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using System.Text;
-using QueryBuilder;
 
-namespace QueryBuilder
+namespace ModelSQLBuilder
 {
     public class ModelBase
     {
