+++ source/ModelSQLBuilder/QueryBuilder.csproj
@@ -1,0 +1,12 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <PackageId>ModelSQLBuilder</PackageId>
+    <Version>1.0.0</Version>
+    <Authors>jzes</Authors>
+    <Company>jzes</Company>
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+  </PropertyGroup>
+
+</Project>
